requirements.txt
pydantic>=2.7
redis>=5.0
pymongo>=4.8
chromadb==0.5.5
numpy>=1.26
python-dotenv>=1.0
openai>=1.52
fastapi>=0.115
uvicorn>=0.30
